Enhanced Randomness through Seeding
Made modifications to the files to infuse seed-based randomness into the Mersenne Twister algorithm, enhancing the random address functionality. A new parameter, ‘seed’, has been introduced in .atp files.
In the absence of an explicitly set seed, ATP will autonomously generate an initial random seed using the random_device() function. However, if a seed is specified, it will be relayed to the random_generator class.
This ‘seed’ parameter can be incorporated within the “RandomDesc” message. Given that “RandomDesc” is utilized for both Sizes and Addresses, it can be employed for either one, both, or none of them.